Update ProductService.java
Updated so that delete works on products existing in users carts, and doesnt delete if the product has been ordered

+++ Clothing4413/src/main/java/com/example/clothing4413/service/ProductService.java
@@ -4,17 +4,25 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
+import com.example.clothing4413.model.CartItem;
+import com.example.clothing4413.model.OrderItem;
 import com.example.clothing4413.model.Product;
 import com.example.clothing4413.model.ProductCategory;
+import com.example.clothing4413.repository.CartItemRepository;
+import com.example.clothing4413.repository.OrderItemRepository;
 import com.example.clothing4413.repository.ProductRepository;
 
 @Service
 public class ProductService {
 
     private final ProductRepository productRepo;
+    private final CartItemRepository cartItemRepo;
+    private final OrderItemRepository orderItemRepo;
 
-    public ProductService(ProductRepository productRepo) {
+    public ProductService(ProductRepository productRepo, CartItemRepository cartItemRepo, OrderItemRepository orderItemRepo) {
         this.productRepo = productRepo;
+        this.cartItemRepo = cartItemRepo;
+        this.orderItemRepo = orderItemRepo;
     }
 
     public List<Product> getAllProducts() {
@@ -75,6 +83,22 @@ public class ProductService {
     }
 
     public void removeProductById(Long id) {
+        Product product = productRepo.findProductById(id);
+        if (product == null) {
+            throw new IllegalArgumentException("Product not found.");
+        }
+
+        //Dont delete if product has been ordered before
+        List<OrderItem> orderItems = orderItemRepo.findByProduct(product);
+        if (!orderItems.isEmpty()) {
+            throw new IllegalStateException("Cannot delete product that has been ordered before.");
+        }
+
+        //Remove product that is being deleted from all carts
+        List<CartItem> cartItems = cartItemRepo.findByProduct(product);
+        cartItemRepo.deleteAll(cartItems);
+
+        //Remove product from product repository
         productRepo.deleteById(id);
     }
